refactor(worldgen): extract engine + hot functions into dedicated modules
Phase 3 of world-builder unification. configBuilder.ts (907 lines) had
nearly the entire builder in one file. Splits it into six focused,
unit-tested modules:

- buildSite.ts: buildFloor + wireStaircases (the shared floor skeleton
  and stairhead-linking, now used by both pyramids and tombs) and
  buildSite (the 3 pyramid floor-shape branches, extracted from
  buildSiteConfigs), plus the per-pyramid randomized resolvers
  (resolveKeyColors, resolveCorridorStraightness, resolvePacking,
  freeWardIndices, buildChestRewards).
- mosaics.ts: computeMosaicPaths.
- fragments.ts: collectSlots, buildPlacementInfos, assignFragments.
- validate.ts: validateRewardCounts, validateDiscovery.

configBuilder.ts shrinks to orchestration: buildPlan, assertChestCapacity,
buildSiteConfigs, buildTombConfigs, buildConfigs. buildTombConfigs now
builds floors via the shared buildFloor + wireStaircases instead of its
own inline exitOrStaircase/entrance bookkeeping.

Dropped in the move: the `decorations` field some floor constructions
spread onto FloorConfig objects. It isn't part of the FloorConfig type,
isn't serialized by serializer.ts, and the app reads a same-named field
from a different, game-layer type (game/siteTypes.ts) that's always
undefined coming from generatedWorld.ts — dead data with no observable
effect either way.

No behavior change — golden guard confirms byte-identical world output;
full suite (694 tests) green.
