+++ src/worldGen/configBuilder.ts
@@ -1,30 +1,23 @@
 import type { Difficulty, FloorConfig, SideSection, SiteConfig, Tier, TreasureReward } from "./types"
-import {
-  PYRAMID_JOURNEYS,
-  TOMB_JOURNEYS,
-  TOMB_SYMBOLS,
-  HIEROGLYPH_REQUIRED,
-  chestEveryFor,
-  chestCountFor,
-} from "./data"
-import { TOMB_PERK_IDS, TREASURE_PERKS } from "../data/treasurePerks"
-import { tableauLevels } from "../data/tableaus"
+import { PYRAMID_JOURNEYS, TOMB_JOURNEYS, HIEROGLYPH_REQUIRED, chestCountFor } from "./data"
+import { TOMB_PERK_IDS } from "../data/treasurePerks"
 import { resolvePyramidConstraintWithProvenance, describeScope } from "./constraintResolver"
 import type { Provenance } from "./constraintResolver"
-import { worldSpec, WORLD_TARGETS } from "./worldSpec"
-import { GLOBAL_DEFAULTS } from "./spec/global"
+import { worldSpec } from "./worldSpec"
 import type {
   PyramidConstraint,
   FloorConstraint,
   RewardHint,
   RewardSpec,
   SideSectionConstraint,
-  SideIntensity,
   PathPuzzlesRange,
 } from "./dsl"
-import { mulberry32 } from "../game/random"
-import { hashStr, hintToReward, rollConsumable, specToReward } from "./rewards"
-import { buildSideSections, pathCountForDensity } from "./sideSections"
+import { hintToReward, specToReward } from "./rewards"
+import { buildSideSections } from "./sideSections"
+import { buildFloor, buildSite, wireStaircases } from "./buildSite"
+import { computeMosaicPaths } from "./mosaics"
+import { assignFragments } from "./fragments"
+import { validateDiscovery, validateRewardCounts } from "./validate"
 
 // ── Ward tier progression ─────────────────────────────────────────────────────
 
@@ -34,28 +27,6 @@ const NEXT_TIER: Record<string, string | null> = {
   expert: "master",
   master: "wizard",
   wizard: null,
-}
-
-// Ward-wing key indices for a tomb, skipping any slot reserved for a tier-unlock or
-// location-key perk (those are spoken for elsewhere) — first `count` remaining indices.
-const freeWardIndices = (tombId: string, count: number): number[] => {
-  const perkIds = TOMB_PERK_IDS[tombId] ?? []
-  const free: number[] = []
-  for (let idx = 0; idx < perkIds.length && free.length < count; idx++) {
-    const perk = TREASURE_PERKS[perkIds[idx]]
-    if (perk?.type !== "tier-unlock" && perk?.type !== "location-key") free.push(idx)
-  }
-  return free
-}
-
-// Secondary tombs that need discovery — primary tomb ID → list of secondary tomb IDs.
-// If a secondary tomb has no mapPiece/locationKey in any authored config, a locationKey
-// is auto-injected as a side section on the primary tomb's last floor.
-const SECONDARY_TOMBS: Record<string, string[]> = {
-  expert_treasure_tomb: ["expert_treasure_tomb_b"],
-  master_treasure_tomb: ["master_treasure_tomb_b"],
-  wizard_treasure_tomb: ["wizard_treasure_tomb_b"],
-  wizard_treasure_tomb_b: ["wizard_treasure_tomb_c"],
 }
 
 // ── Path puzzle scaling ───────────────────────────────────────────────────────
@@ -71,151 +42,6 @@ const interpolatePathPuzzles = (range: PathPuzzlesRange, i: number, total: numbe
 
 const resolvePathPuzzles = (value: number | PathPuzzlesRange, i: number, total: number): number =>
   isPathPuzzlesRange(value) ? interpolatePathPuzzles(value, i, total) : value
-
-// ── Chest rewards ─────────────────────────────────────────────────────────────
-
-const buildChestRewards = (
-  journeyId: string,
-  slotOffset: number,
-  pathPuzzles: number,
-  rates: { bandage: number; oil: number; trapTool: number } = GLOBAL_DEFAULTS.consumableRates
-): TreasureReward[] => {
-  const count = chestCountFor(pathPuzzles)
-  return Array.from({ length: count }, (_, i) => ({
-    type: "consumable" as const,
-    consumable: rollConsumable(`${journeyId}:consumable:${slotOffset + i}`, rates),
-  }))
-}
-
-// ── Mosaic path distribution ──────────────────────────────────────────────────
-
-const INTENSITY_PATHS: Record<SideIntensity, number> = { none: 0, low: 1, medium: 2, dense: 4 }
-
-// Resolves the effective keyColors for a pyramid, honoring (in priority order):
-// a literal keyColorsRange roll, a literal keyColors, then a sharedKeyChance roll — a hit
-// resolves to 1 (one key, every gated door); a miss resolves to 5 (mostly-dedicated keys),
-// since the buildSideSections default of "no keyColors set" already means 1 color and
-// would make a miss indistinguishable from a hit.
-const resolveKeyColors = (
-  constraint: PyramidConstraint,
-  journeyId: string,
-  pyramidIndex: number
-): number | undefined => {
-  const rand = mulberry32(hashStr(`${journeyId}:${pyramidIndex}:keyColors`))
-  if (constraint.keyColorsRange) {
-    const { min, max } = constraint.keyColorsRange
-    return min + Math.floor(rand() * (max - min + 1))
-  }
-  if (constraint.keyColors !== undefined) return constraint.keyColors
-  if (constraint.sharedKeyChance !== undefined) return rand() < constraint.sharedKeyChance ? 1 : 5
-  return undefined
-}
-
-// Resolves an authored literal, or a chance-rolled "hit" value, or undefined (builder default).
-const resolveChanceValue = (
-  literal: number | undefined,
-  chance: number | undefined,
-  hitValue: number,
-  journeyId: string,
-  pyramidIndex: number,
-  tag: string
-): number | undefined => {
-  if (literal !== undefined) return literal
-  if (chance === undefined) return undefined
-  const rand = mulberry32(hashStr(`${journeyId}:${pyramidIndex}:${tag}`))
-  return rand() < chance ? hitValue : undefined
-}
-
-const resolveCorridorStraightness = (constraint: PyramidConstraint, journeyId: string, pyramidIndex: number) =>
-  resolveChanceValue(
-    constraint.corridorStraightness,
-    constraint.windyChance,
-    constraint.windyStraightness ?? GLOBAL_DEFAULTS.windyStraightness,
-    journeyId,
-    pyramidIndex,
-    "windy"
-  )
-
-const resolvePacking = (constraint: PyramidConstraint, journeyId: string, pyramidIndex: number) =>
-  resolveChanceValue(
-    constraint.packing,
-    constraint.packingChance,
-    constraint.packingWhenHit ?? GLOBAL_DEFAULTS.packingWhenHit,
-    journeyId,
-    pyramidIndex,
-    "packing"
-  )
-
-const computeMosaicPaths = (plan: PyramidPlan[]): Map<string, number> => {
-  let committed = 0
-  for (const p of plan) {
-    if (p.constraint.mainEndReward === "mosaicPiece") committed++
-  }
-
-  const explicitPaths = new Map<string, number>()
-  const autoCandidates: PyramidPlan[] = []
-  let explicitTotal = 0
-
-  for (const p of plan) {
-    const key = `${p.journeyId}:${p.pyramidIndex}`
-    // Multi-floor pyramids with explicit floors[] are fully specified — exclude from auto-distribution
-    if (p.constraint.floors?.length) {
-      for (const floor of p.constraint.floors) {
-        const floorSd = floor?.sideSections
-        if (Array.isArray(floorSd)) committed += floorSd.filter(s => s.endReward === "mosaicPiece").length
-      }
-      explicitPaths.set(key, 0)
-      continue
-    }
-    const sd = p.constraint.sideSections
-    if (typeof sd === "string") {
-      // SideIntensity → all side paths are mosaic, not an auto-candidate
-      const count = INTENSITY_PATHS[sd as SideIntensity] ?? 0
-      explicitPaths.set(key, count)
-      explicitTotal += count
-    } else if (typeof sd === "number") {
-      explicitPaths.set(key, sd)
-      explicitTotal += sd
-    } else {
-      // Array or undefined → auto-candidate; count explicitly specified mosaicPiece sections
-      if (Array.isArray(sd)) {
-        committed += sd.filter(s => s.endReward === "mosaicPiece").length
-      }
-      // Count mosaic paths from sidePaths/hiddenPaths declarations — those pyramids leave auto-pool.
-      // mosaicPathCount is set to 0 so buildSideSections skips the auto-mosaic loop;
-      // the declared hidden mosaics are built directly from constraints.
-      const allDeclared = [...(p.constraint.sidePaths ?? []), ...(p.constraint.hiddenPaths ?? [])]
-      const declaredMosaics = allDeclared.filter(e => e.end === "mosaic")
-      if (declaredMosaics.length > 0) {
-        const count = declaredMosaics.reduce(
-          (sum, e) => sum + pathCountForDensity(e.density, p.journeyId, p.pyramidIndex),
-          0
-        )
-        committed += count
-        explicitPaths.set(key, 0) // buildSideSections handles these via declaredHiddenPaths
-      } else {
-        autoCandidates.push(p)
-      }
-    }
-  }
-
-  const remaining = WORLD_TARGETS.mosaicPieceRewards - committed - explicitTotal
-  const result = new Map(explicitPaths)
-
-  if (remaining > 0 && autoCandidates.length > 0) {
-    const sorted = [...autoCandidates].sort((a, b) => b.pathPuzzles - a.pathPuzzles)
-    for (let rem = remaining, i = 0; rem > 0; rem--, i++) {
-      const p = sorted[i % sorted.length]
-      const key = `${p.journeyId}:${p.pyramidIndex}`
-      result.set(key, (result.get(key) ?? 0) + 1)
-    }
-  }
-
-  return result
-}
-
-// ── Side sections ─────────────────────────────────────────────────────────────
-// buildSideSections itself lives in ./sideSections — shared by pyramids and tombs.
 
 // ── Phase 1: Build initial plan ───────────────────────────────────────────────
 
@@ -325,246 +151,29 @@ const buildSiteConfigs = (plan: PyramidPlan[]): Record<string, SiteConfig[]> => 
       const { pyramidIndex: i, pathPuzzles: pp, constraint } = p
       const difficulty: Difficulty = constraint.difficulty ?? "expert"
 
-      const hasMapPieceBranch = i === mapPiecePyramid && tier !== "starter"
-      const hasWardGate = i >= Math.ceil(levelCount / 2) && nextTier !== null
-
-      const mainEndReward: TreasureReward = constraint.mainEndReward
-        ? specToReward(constraint.mainEndReward, tier)
-        : { type: "fragmentSlot" }
-
-      if (constraint.floors?.length) {
-        // Multi-floor: build one FloorConfig per floors[] entry
-        const floorConfigs: FloorConfig[] = []
-        for (let fi = 0; fi < constraint.floors.length; fi++) {
-          const fc = constraint.floors[fi] ?? {}
-          const floorPP = typeof fc.pathPuzzles === "number" ? fc.pathPuzzles : pp
-          const floorDiff: Difficulty = fc.difficulty ?? difficulty
-          const isLast = fi === constraint.floors.length - 1
-          const floorSections = Array.isArray(fc.sideSections) ? fc.sideSections : undefined
-          const floorSideSections = buildSideSections({
-            tier,
-            difficulty: floorDiff,
-            resolveReward: spec => specToReward(spec, tier),
-            journeyId,
-            constraintSections: floorSections,
-          })
-          const floorChests = buildChestRewards(journeyId, chestOffset, floorPP, constraint.consumableRates)
-          chestOffset += chestCountFor(floorPP)
-          const floorStraightness = fc.corridorStraightness ?? resolveCorridorStraightness(constraint, journeyId, i)
-          const floorPacking = fc.packing ?? resolvePacking(constraint, journeyId, i)
-          floorConfigs.push({
-            pathPuzzles: floorPP,
-            chestEvery: chestEveryFor(floorPP),
-            difficulty: floorDiff,
-            end: "treasure",
-            exitOrStaircase: "exit",
-            sideSections: floorSideSections,
-            ...(isLast ? { mainEndReward } : {}),
-            ...(floorChests.length > 0 ? { chestRewards: floorChests } : {}),
-            ...(fc.decorations?.length ? { decorations: fc.decorations } : {}),
-            ...(floorStraightness !== undefined ? { corridorStraightness: floorStraightness } : {}),
-            ...(floorPacking !== undefined ? { packing: floorPacking } : {}),
-          } satisfies FloorConfig)
-        }
-        // Wire each floor's side-path stairhead to the entrance of the next floor.
-        for (let fi = 0; fi < floorConfigs.length - 1; fi++) {
-          const stairSection = floorConfigs[fi].sideSections.find(s => typeof s.end === "object")
-          if (stairSection && typeof stairSection.end === "object") {
-            floorConfigs[fi + 1].entrance = { stairId: stairSection.end.stairId }
-          }
-        }
-        pyramidConfigs.push(floorConfigs)
-      } else if (
-        (constraint.mainFloors ?? GLOBAL_DEFAULTS.mainFloors) > 1 ||
-        (constraint.wardWings ?? GLOBAL_DEFAULTS.wardWings) > 0
-      ) {
-        // Auto multi-floor: `mainFloors` plain main-path floors (only the last one carries
-        // the pyramid's usual side content), then `wardWings` bonus floors branching off
-        // that last main floor, each behind its own ward-key gate from this tier's own tomb.
-        const mainFloors = constraint.mainFloors ?? GLOBAL_DEFAULTS.mainFloors
-        const wardWings = constraint.wardWings ?? GLOBAL_DEFAULTS.wardWings
-        const floorConfigs: FloorConfig[] = []
-
-        for (let fi = 0; fi < mainFloors; fi++) {
-          if (fi < mainFloors - 1) {
-            floorConfigs.push({
-              pathPuzzles: pp,
-              chestEvery: 0,
-              difficulty,
-              end: "treasure",
-              exitOrStaircase: "exit",
-              sideSections: [],
-            })
-            continue
-          }
-          const constraintSections = Array.isArray(constraint.sideSections) ? constraint.sideSections : undefined
-          const mosaicPathCount = mosaicPaths.get(`${journeyId}:${i}`) ?? 0
-          const sideSections = buildSideSections({
-            tier,
-            difficulty,
-            resolveReward: spec => specToReward(spec, tier),
-            journeyId,
-            constraintSections,
-            hasMapPieceBranch,
-            hasWardGate,
-            nextTier,
-            mosaicPathCount,
-            mainPathPuzzles: pp,
-            keyDensity: constraint.keyDensity,
-            keyColors: resolveKeyColors(constraint, journeyId, i),
-            pyramidIndex: i,
-            declaredSidePaths: constraint.sidePaths,
-            declaredHiddenPaths: constraint.hiddenPaths,
-          })
-          const chestRewards = buildChestRewards(journeyId, chestOffset, pp, constraint.consumableRates)
-          chestOffset += chestCountFor(pp)
-          const straightness = resolveCorridorStraightness(constraint, journeyId, i)
-          const packing = resolvePacking(constraint, journeyId, i)
-          floorConfigs.push({
-            pathPuzzles: pp,
-            chestEvery: chestEveryFor(pp),
-            difficulty,
-            end: "treasure",
-            exitOrStaircase: "exit",
-            sideSections,
-            mainEndReward,
-            chestRewards,
-            ...(constraint.consumableDensity !== undefined ? { consumableDensity: constraint.consumableDensity } : {}),
-            ...(straightness !== undefined ? { corridorStraightness: straightness } : {}),
-            ...(packing !== undefined ? { packing } : {}),
-          } satisfies FloorConfig)
-        }
-
-        // Wire main-floor stairheads sequentially (floor N's exit → floor N+1's entrance).
-        for (let fi = 0; fi < floorConfigs.length - 1; fi++) {
-          const stairId = `${journeyId}:p${i}:main${fi}`
-          floorConfigs[fi].exitOrStaircase = { stairId }
-          floorConfigs[fi + 1].entrance = { stairId }
-        }
-
-        if (wardWings > 0) {
-          const tombId = `${tier}_treasure_tomb`
-          const wingIndices = freeWardIndices(tombId, wardWings)
-          const lastMain = floorConfigs[floorConfigs.length - 1]
-          for (let w = 0; w < wingIndices.length; w++) {
-            const wingStairId = `${journeyId}:p${i}:wing${w}`
-            lastMain.sideSections = [
-              ...lastMain.sideSections,
-              {
-                pathPuzzles: 1,
-                difficulty,
-                end: { stairId: wingStairId },
-                gate: { type: "tomb-key", wardKeyId: TOMB_PERK_IDS[tombId][wingIndices[w]] },
-              },
-            ]
-            floorConfigs.push({
-              pathPuzzles: pp,
-              chestEvery: chestEveryFor(pp),
-              difficulty,
-              end: "treasure",
-              exitOrStaircase: "exit",
-              entrance: { stairId: wingStairId },
-              sideSections: [],
-              mainEndReward: { type: "hieroglyphs" },
-            })
-          }
-        }
-
-        pyramidConfigs.push(floorConfigs)
-      } else {
-        const constraintSections = Array.isArray(constraint.sideSections) ? constraint.sideSections : undefined
-        const mosaicPathCount = mosaicPaths.get(`${journeyId}:${i}`) ?? 0
-        const sideSections = buildSideSections({
-          tier,
-          difficulty,
-          resolveReward: spec => specToReward(spec, tier),
-          journeyId,
-          constraintSections,
-          hasMapPieceBranch,
-          hasWardGate,
-          nextTier,
-          mosaicPathCount,
-          mainPathPuzzles: pp,
-          keyDensity: constraint.keyDensity,
-          keyColors: resolveKeyColors(constraint, journeyId, i),
-          pyramidIndex: i,
-          declaredSidePaths: constraint.sidePaths,
-          declaredHiddenPaths: constraint.hiddenPaths,
-        })
-        const chestRewards = buildChestRewards(journeyId, chestOffset, pp, constraint.consumableRates)
-        chestOffset += chestCountFor(pp)
-        const consumableDensity = constraint.consumableDensity
-        const straightness = resolveCorridorStraightness(constraint, journeyId, i)
-        const packing = resolvePacking(constraint, journeyId, i)
-        pyramidConfigs.push([
-          {
-            pathPuzzles: pp,
-            chestEvery: chestEveryFor(pp),
-            difficulty,
-            end: "treasure",
-            exitOrStaircase: "exit",
-            sideSections,
-            mainEndReward,
-            chestRewards,
-            ...(consumableDensity !== undefined ? { consumableDensity } : {}),
-            ...(straightness !== undefined ? { corridorStraightness: straightness } : {}),
-            ...(packing !== undefined ? { packing } : {}),
-          } satisfies FloorConfig,
-        ])
-      }
+      const { floors, chestOffset: nextChestOffset } = buildSite({
+        journeyId,
+        tier,
+        pyramidIndex: i,
+        pathPuzzles: pp,
+        constraint,
+        difficulty,
+        hasMapPieceBranch: i === mapPiecePyramid && tier !== "starter",
+        hasWardGate: i >= Math.ceil(levelCount / 2) && nextTier !== null,
+        nextTier,
+        mosaicPathCount: mosaicPaths.get(`${journeyId}:${i}`) ?? 0,
+        chestOffset,
+        resolveReward: spec => specToReward(spec, tier),
+        resolveMainEndReward: spec => specToReward(spec, tier),
+      })
+      chestOffset = nextChestOffset
+      pyramidConfigs.push(floors)
     }
 
     configs[journeyId] = pyramidConfigs
   }
 
   return configs
-}
-
-// ── Phase 5: Validate structural rewards ──────────────────────────────────────
-
-const KNOWN_JOURNEY_IDS = new Set([...PYRAMID_JOURNEYS.map(j => j.id), ...TOMB_JOURNEYS.map(j => j.id)])
-
-const validateRewardCounts = (configs: Record<string, SiteConfig[]>): void => {
-  let mapPieces = 0
-  let mosaicPieces = 0
-  const unknownTombIds: string[] = []
-
-  const checkReward = (r: TreasureReward | undefined) => {
-    if (!r) return
-    if (r.type === "mapPiece") {
-      mapPieces++
-      if (!KNOWN_JOURNEY_IDS.has(r.tombId)) unknownTombIds.push(r.tombId)
-    }
-    if (r.type === "mosaicPiece") mosaicPieces++
-  }
-
-  for (const [siteId, siteConfigs] of Object.entries(configs)) {
-    for (const floors of siteConfigs) {
-      for (let fi = 0; fi < floors.length; fi++) {
-        const floor = floors[fi]
-        const isLast = fi === floors.length - 1
-        if (isLast && floor.exitOrStaircase !== "exit")
-          throw new Error(
-            `[worldSpec] Site "${siteId}" last floor has exitOrStaircase="${floor.exitOrStaircase}", expected "exit"`
-          )
-        checkReward(floor.mainEndReward)
-        for (const r of floor.chestRewards ?? []) checkReward(r)
-        for (const s of floor.sideSections) {
-          checkReward(s.endReward)
-          for (const sub of s.sideSections ?? []) checkReward(sub.endReward)
-        }
-      }
-    }
-  }
-
-  if (unknownTombIds.length > 0)
-    throw new Error(
-      `[worldSpec] mapPiece rewards reference unknown journey IDs: ${[...new Set(unknownTombIds)].join(", ")}`
-    )
-  if (mapPieces !== WORLD_TARGETS.mapPieceRewards)
-    throw new Error(`[worldSpec] Expected ${WORLD_TARGETS.mapPieceRewards} map pieces, got ${mapPieces}`)
-  if (mosaicPieces !== WORLD_TARGETS.mosaicPieceRewards)
-    throw new Error(`[worldSpec] Expected ${WORLD_TARGETS.mosaicPieceRewards} mosaic pieces, got ${mosaicPieces}`)
 }
 
 // ── Tomb configs ──────────────────────────────────────────────────────────────
@@ -595,7 +204,7 @@ const buildTombConfigs = (): Record<string, SiteConfig[]> => {
       return undefined
     }
 
-    const floors: SiteConfig = Array.from({ length: levelCount }, (_, i) => {
+    const floors: FloorConfig[] = Array.from({ length: levelCount }, (_, i) => {
       const isLast = i === levelCount - 1
       const authored = authoredFloors?.[i]
 
@@ -620,266 +229,23 @@ const buildTombConfigs = (): Record<string, SiteConfig[]> => {
       const straightness = authored?.corridorStraightness ?? constraint.corridorStraightness
       const packing = authored?.packing ?? constraint.packing
 
-      return {
+      return buildFloor({
         pathPuzzles: isLast && hasCroc ? 2 : 1,
         chestEvery: 0,
         difficulty,
-        end: "treasure" as const,
-        exitOrStaircase: isLast ? ("exit" as const) : { stairId: `${tomb.id}:floor${i}` },
-        ...(i > 0 ? { entrance: { stairId: `${tomb.id}:floor${i - 1}` } } : {}),
         sideSections,
         puzzleFamily,
-        ...(isLast && hasCroc ? { lastMainPuzzleFamily: "crocodile" as const } : {}),
-        ...(mainEndReward ? { mainEndReward } : {}),
-        ...(authored?.decorations?.length ? { decorations: authored.decorations } : {}),
-        ...(straightness !== undefined ? { corridorStraightness: straightness } : {}),
-        ...(packing !== undefined ? { packing } : {}),
-      }
+        lastMainPuzzleFamily: isLast && hasCroc ? "crocodile" : undefined,
+        mainEndReward,
+        corridorStraightness: straightness,
+        packing,
+      })
     })
 
+    wireStaircases(floors, fi => `${tomb.id}:floor${fi}`)
     configs[tomb.id] = [floors]
   }
   return configs
-}
-
-// ── Phase 7: Validate discovery graph ────────────────────────────────────────
-
-// Collect all tombIds that have a mapPiece reward in any config OTHER than their own site
-const collectDiscoveredBy = (configs: Record<string, SiteConfig[]>): Map<string, Set<string>> => {
-  const discovered = new Map<string, Set<string>>()
-  for (const [siteId, siteConfigs] of Object.entries(configs)) {
-    for (const floors of siteConfigs) {
-      for (const floor of floors) {
-        const checkReward = (r: TreasureReward | undefined) => {
-          if (r?.type !== "mapPiece" || r.tombId === siteId) return
-          const set = discovered.get(r.tombId) ?? new Set()
-          set.add(siteId)
-          discovered.set(r.tombId, set)
-        }
-        checkReward(floor.mainEndReward)
-        for (const s of floor.sideSections) {
-          checkReward(s.endReward)
-          for (const sub of s.sideSections ?? []) checkReward(sub.endReward)
-        }
-        for (const r of floor.chestRewards ?? []) checkReward(r)
-      }
-    }
-  }
-  return discovered
-}
-
-// Validate that every secondary tomb has a mapPiece reward reachable before it's needed.
-// Throws with a clear message listing any unreachable secondary tombs (missing or circular).
-const validateDiscovery = (allConfigs: Record<string, SiteConfig[]>): void => {
-  const allSecondary = new Set(Object.values(SECONDARY_TOMBS).flat())
-  const discoveredBy = collectDiscoveredBy(allConfigs)
-
-  // BFS: start from non-secondary sites (auto-discovered), expand when mapPiece host is reachable
-  const reachable = new Set(Object.keys(allConfigs).filter(id => !allSecondary.has(id)))
-  let changed = true
-  while (changed) {
-    changed = false
-    for (const secId of allSecondary) {
-      if (reachable.has(secId)) continue
-      const hosts = discoveredBy.get(secId)
-      if (hosts && [...hosts].some(h => reachable.has(h))) {
-        reachable.add(secId)
-        changed = true
-      }
-    }
-  }
-
-  const unreachable = [...allSecondary].filter(id => !reachable.has(id))
-  if (unreachable.length > 0) {
-    throw new Error(
-      `[worldSpec] Unsolvable discovery graph — these secondary tombs are unreachable:\n` +
-        unreachable.map(id => `  - ${id} (no mapPiece found in a reachable site)`).join("\n")
-    )
-  }
-}
-
-// ── Phase 9f: Fragment assignment ────────────────────────────────────────────
-
-const TIERS: Tier[] = ["starter", "junior", "expert", "master", "wizard"]
-
-export type SlotRef = {
-  journeyId: string
-  tier: Tier
-  journeyOrderIndex: number
-  wardKeys: string[]
-  isPlaceholder: boolean
-  assign: (r: TreasureReward) => void
-}
-
-export type HieroglyphPlacementInfo = {
-  hieroglyphId: string
-  tier: Tier
-  preferredWardKeys: string[]
-  required: number
-}
-
-export const collectSlots = (allConfigs: Record<string, SiteConfig[]>): SlotRef[] => {
-  const slots: SlotRef[] = []
-
-  for (const [journeyId, siteConfigs] of Object.entries(allConfigs)) {
-    const journey = PYRAMID_JOURNEYS.find(j => j.id === journeyId)
-    if (!journey) continue
-
-    const tier = journey.tier as Tier
-    const journeyOrderIndex = PYRAMID_JOURNEYS.indexOf(journey)
-
-    const addSlot = (wardKeys: string[], isPlaceholder: boolean, assign: (r: TreasureReward) => void) =>
-      slots.push({ journeyId, tier, journeyOrderIndex, wardKeys, isPlaceholder, assign })
-
-    for (const floors of siteConfigs) {
-      for (const floor of floors) {
-        if (floor.mainEndReward?.type === "fragmentSlot") {
-          const f = floor
-          addSlot([], true, r => {
-            f.mainEndReward = r
-          })
-        }
-        for (const section of floor.sideSections) {
-          const sWardKeys = section.gate?.type === "tomb-key" ? [section.gate.wardKeyId] : []
-          if (section.endReward?.type === "fragmentSlot") {
-            const s = section
-            addSlot(sWardKeys, true, r => {
-              s.endReward = r
-            })
-          } else if (section.gate?.type === "tomb-key" && !section.endReward) {
-            const s = section
-            addSlot(sWardKeys, false, r => {
-              s.endReward = r
-            })
-          }
-          for (const sub of section.sideSections ?? []) {
-            const subWardKeys = [...sWardKeys, ...(sub.gate?.type === "tomb-key" ? [sub.gate.wardKeyId] : [])]
-            if (sub.endReward?.type === "fragmentSlot") {
-              const ss = sub
-              addSlot(subWardKeys, true, r => {
-                ss.endReward = r
-              })
-            } else if (sub.gate?.type === "tomb-key" && !sub.endReward) {
-              const ss = sub
-              addSlot(subWardKeys, false, r => {
-                ss.endReward = r
-              })
-            }
-          }
-        }
-      }
-    }
-  }
-
-  return slots
-}
-
-export const buildPlacementInfos = (): HieroglyphPlacementInfo[] => {
-  const infos: HieroglyphPlacementInfo[] = []
-  const seen = new Set<string>()
-
-  for (const tier of TIERS) {
-    const tombId = `${tier}_treasure_tomb`
-    const tombPerkIds = TOMB_PERK_IDS[tombId] ?? []
-
-    for (const hieroglyphId of TOMB_SYMBOLS[tier as Tier]) {
-      if (seen.has(hieroglyphId)) continue
-      seen.add(hieroglyphId)
-
-      const firstRunNumber = tableauLevels
-        .filter(t => t.tombJourneyId === tombId && t.inventoryIds.includes(hieroglyphId))
-        .reduce((min, t) => Math.min(min, t.runNumber), Infinity)
-
-      const runNumber = isFinite(firstRunNumber) ? firstRunNumber : 1
-      // Ward keys earned after completing runs 1..(runNumber-1) gate the preferred slots.
-      // run 1 → no wards needed; run 2 → tombPerkIds[0]; run 3 → tombPerkIds[0..1]; etc.
-      const preferredWardKeys = tombPerkIds.slice(0, runNumber - 1)
-
-      infos.push({
-        hieroglyphId,
-        tier: tier as Tier,
-        preferredWardKeys,
-        required: HIEROGLYPH_REQUIRED[hieroglyphId] ?? 2,
-      })
-    }
-  }
-
-  return infos
-}
-
-const assignFragments = (allConfigs: Record<string, SiteConfig[]>): void => {
-  const slots = collectSlots(allConfigs)
-  const infos = buildPlacementInfos()
-  const available = [...slots]
-
-  const placedInJourney = new Map<string, Set<string>>()
-  for (const j of PYRAMID_JOURNEYS) placedInJourney.set(j.id, new Set())
-
-  let totalPlaced = 0
-
-  for (const info of infos) {
-    const needed = info.required
-    let placed = 0
-
-    // Pools in priority order:
-    // 0 — tier-matching slots behind preferred ward keys (run 2+ fragments go here first)
-    // 1 — tier-matching open slots (no ward)
-    // 2 — any remaining slots (cross-tier fallback)
-    const pools = [
-      available.filter(
-        s =>
-          s.tier === info.tier &&
-          info.preferredWardKeys.length > 0 &&
-          s.wardKeys.some(k => info.preferredWardKeys.includes(k))
-      ),
-      available.filter(s => s.tier === info.tier && s.wardKeys.length === 0),
-      available.filter(s => s.tier !== info.tier),
-    ]
-
-    for (const pool of pools) {
-      if (placed >= needed) break
-
-      // First pass: respect 1-per-journey
-      for (const slot of [...pool]) {
-        if (placed >= needed) break
-        const idx = available.indexOf(slot)
-        if (idx === -1) continue
-        if (placedInJourney.get(slot.journeyId)?.has(info.hieroglyphId)) continue
-        slot.assign({ type: "hieroglyphFragment", hieroglyphId: info.hieroglyphId })
-        placedInJourney.get(slot.journeyId)!.add(info.hieroglyphId)
-        available.splice(idx, 1)
-        placed++
-      }
-
-      if (placed >= needed) break
-
-      // Second pass: relax 1-per-journey if pool exhausted
-      for (const slot of [...pool]) {
-        if (placed >= needed) break
-        const idx = available.indexOf(slot)
-        if (idx === -1) continue
-        slot.assign({ type: "hieroglyphFragment", hieroglyphId: info.hieroglyphId })
-        available.splice(idx, 1)
-        placed++
-      }
-    }
-
-    totalPlaced += placed
-    if (placed < needed) {
-      console.warn(`  ⚠ ${info.hieroglyphId} (${info.tier}): placed ${placed}/${needed} — not enough fragment slots`)
-    }
-  }
-
-  // Fill remaining placeholder slots with consumables
-  const rates = GLOBAL_DEFAULTS.consumableRates
-  let fallbackIdx = 0
-  for (const slot of available) {
-    if (!slot.isPlaceholder) continue
-    const consumable = rollConsumable(`${slot.journeyId}:fragment-fallback:${fallbackIdx++}`, rates)
-    slot.assign({ type: "consumable", consumable })
-  }
-
-  console.log(`  ✓ Fragment assignment: ${totalPlaced} fragments placed`)
 }
 
 // ── Main entry point ──────────────────────────────────────────────────────────
